Update curl command line in README installation instructions.
The change adds the -L switch, which asks curl to follow redirects 
(needed to follow the redirect to HTTPS), and the -k switch to ignore 
certificate errors. I'm not sure if the latter is a good idea, but it's 
necessary when using curl on CygWin. Note that not including the -k 
switch may cause curl to fail the download silently, with no indication 
of what the problem is, due to the -s switch.